// vehicle-service-management/src/app/core/api/generated/types.gen.ts
export type CreateVehicleDto = {
  plateNumber: string;
  brand: string;
  model: string;
  year: number | string;
  color: string;
  vin?: null | string;
  currentMileage: number | string;
  customerId?: null | number | string;
};

export type CustomerDto = {
  id: number | string;
  fullName: string;
  contactNumber: string;
  email: string;
  address: string;
  createdAt: string;
  updatedAt: string;
  isArchived: boolean;
  rowVersion: string;
};

export type CustomerSummaryDto = {
  id: number | string;
  fullName: string;
  isArchived: boolean;
};
// ...
export type VehicleDto = {
  id: number | string;
  plateNumber: string;
  brand: string;
  model: string;
  year: number | string;
  color: string;
  vin?: null | string;
  currentMileage: number | string;
  customerId?: null | number | string;
  customer?: null | CustomerSummaryDto;
  createdAt: string;
  updatedAt: string;
  isArchived: boolean;
  rowVersion: string;
};